requirements.txt
altair==4.2.0
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
black==22.8.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
codecov==2.1.12
commonmark==0.9.1
coverage==6.4.4
cryptography==38.0.1
db-dtypes==1.0.4
decorator==5.1.1
dynaconf==3.1.11
ecdsa==0.18.0
entrypoints==0.4
fastapi==0.85.0
flake8==5.0.4
ghp-import==2.1.0
gitchangelog==3.0.4
gitdb==4.0.9
GitPython==3.1.27
google-api-core
google-auth
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet==1.1.3
grpcio
grpcio-status
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==4.16.0
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.0
mypy==0.981
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
pandas==1.5.0
passlib==1.7.4
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
proto-plus
protobuf
psycopg2-binary==2.9.3
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b3
pyflakes==2.5.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-cov==3.0.0
pytest-picked==0.4.6
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.1
rich==12.5.1
rsa==4.9
seaborn==0.12.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.20.4
streamlit==1.13.0
toml==0.10.2
tomli==2.0.1
toolz
tornado
typer
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
validators
watchdog
zipp
